Remove src/control and use src/stagbl.c
