--- src/control/local.mk
@@ -1,3 +1,0 @@
-libstagbl-y.c += $(call thisdir,\
-  stagbl.c\
-)
