+++ src/local.mk
@@ -2,4 +2,4 @@ libstagbl-y.c += $(call thisdir,\
   stagbl.c\
 )
 
-include $(call incsubdirs,control core stokes)
+include $(call incsubdirs,core stokes)
